fix: cast turbo text tokens to int64 for the repetition-penalty gather
The upstream cfg path normalizes token dtype inside t3.inference, but
inference_turbo does not, and its RepetitionPenaltyLogitsProcessor gather
rejects int32 indices on MPS. Verified with the chatterbox-turbo hardware
smoke test.

+++ src/bilbo_tts/tts/chatterbox.py
@@ -359,10 +359,16 @@ class ChatterboxTtsEngine:
         helpers = _import_turbo_helpers()
         raw: Any = model
         torch: Any = dependencies.torch
-        text_tokens: Any = raw.tokenizer.text_to_tokens(
-            helpers.punc_norm(spoken_text),
-            language_id="it",
-        ).to("mps")
+        # Unlike the cfg path, inference_turbo does not normalize token dtype,
+        # and its repetition-penalty gather requires int64 indices.
+        text_tokens: Any = (
+            raw.tokenizer.text_to_tokens(
+                helpers.punc_norm(spoken_text),
+                language_id="it",
+            )
+            .to("mps")
+            .to(torch.long)
+        )
         text_tokens = helpers.pad(text_tokens, (1, 0), value=raw.t3.hp.start_text_token)
         text_tokens = helpers.pad(text_tokens, (0, 1), value=raw.t3.hp.stop_text_token)
         with torch.inference_mode():
